Manual fixes of source code: 1. Fixed errors for the autoconverted code 2. Implemented missing functionality in the Sharpen project 3. Readme + some documentation 4. Appveyor config 5. Doxygen config

+++ Com.Drew.Tests/Com/Drew/Lang/SequentialAccessTestBase.cs
@@ -92,7 +92,7 @@ namespace Com.Drew.Lang
 			reader = CreateReader(buffer);
 			reader.SetMotorolaByteOrder(false);
 			Sharpen.Tests.AreEqual((short)unchecked((int)(0x0100)), reader.GetInt16());
-			Sharpen.Tests.AreEqual((short)unchecked((int)(0xFF7F)), reader.GetInt16());
+			Sharpen.Tests.AreEqual(unchecked((short)(0xFF7F)), reader.GetInt16());
 		}
 
 		/// <exception cref="System.IO.IOException"/>
@@ -128,8 +128,10 @@ namespace Com.Drew.Lang
 		[NUnit.Framework.Test]
 		public virtual void TestGetInt32()
 		{
-			Sharpen.Tests.AreEqual(-1, CreateReader(new sbyte[] { unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))) }).GetInt32());
-			sbyte[] buffer = new sbyte[] { unchecked((int)(0x00)), unchecked((int)(0x01)), unchecked((int)(0x02)), unchecked((int)(0x03)), unchecked((int)(0x04)), unchecked((int)(0x05)), unchecked((int)(0x06)), unchecked((int)(0x07)) };
+			Sharpen.Tests.AreEqual(-1, CreateReader(new sbyte[] { unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte
+				)unchecked((int)(0xff))) }).GetInt32());
+			sbyte[] buffer = new sbyte[] { unchecked((int)(0x00)), unchecked((int)(0x01)), unchecked((int)(0x02)), unchecked((int)(0x03)), unchecked((int)(0x04)), unchecked((int)(0x05)), unchecked((int)(0x06)), unchecked(
+				(int)(0x07)) };
 			SequentialReader reader = CreateReader(buffer);
 			Sharpen.Tests.AreEqual(unchecked((int)(0x00010203)), reader.GetInt32());
 			Sharpen.Tests.AreEqual(unchecked((int)(0x04050607)), reader.GetInt32());
@@ -143,9 +145,10 @@ namespace Com.Drew.Lang
 		[NUnit.Framework.Test]
 		public virtual void TestGetUInt32()
 		{
-			Sharpen.Tests.AreEqual(4294967295L, CreateReader(new sbyte[] { unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))) }).GetUInt32
-				());
-			sbyte[] buffer = new sbyte[] { unchecked((sbyte)unchecked((int)(0xFF))), unchecked((int)(0x00)), unchecked((int)(0x01)), unchecked((int)(0x02)), unchecked((int)(0x03)), unchecked((int)(0x04)), unchecked((int)(0x05)), unchecked((int)(0x06)) };
+			Sharpen.Tests.AreEqual(4294967295L, CreateReader(new sbyte[] { unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xff))), unchecked(
+				(sbyte)unchecked((int)(0xff))) }).GetUInt32());
+			sbyte[] buffer = new sbyte[] { unchecked((sbyte)unchecked((int)(0xFF))), unchecked((int)(0x00)), unchecked((int)(0x01)), unchecked((int)(0x02)), unchecked((int)(0x03)), unchecked((int)(0x04)), unchecked(
+				(int)(0x05)), unchecked((int)(0x06)) };
 			SequentialReader reader = CreateReader(buffer);
 			Sharpen.Tests.AreEqual(unchecked((long)(0xFF000102L)), reader.GetUInt32());
 			Sharpen.Tests.AreEqual(unchecked((long)(0x03040506L)), reader.GetUInt32());
@@ -175,8 +178,8 @@ namespace Com.Drew.Lang
 		[NUnit.Framework.Test]
 		public virtual void TestGetInt64()
 		{
-			sbyte[] buffer = new sbyte[] { unchecked((sbyte)unchecked((int)(0xFF))), unchecked((int)(0x00)), unchecked((int)(0x01)), unchecked((int)(0x02)), unchecked((int)(0x03)), unchecked((int)(0x04)), unchecked((int)(0x05)), unchecked((int)(0x06)), 
-				unchecked((int)(0x07)) };
+			sbyte[] buffer = new sbyte[] { unchecked((sbyte)unchecked((int)(0xFF))), unchecked((int)(0x00)), unchecked((int)(0x01)), unchecked((int)(0x02)), unchecked((int)(0x03)), unchecked((int)(0x04)), unchecked(
+				(int)(0x05)), unchecked((int)(0x06)), unchecked((int)(0x07)) };
 			SequentialReader reader = CreateReader(buffer);
 			Sharpen.Tests.AreEqual(unchecked((long)(0xFF00010203040506L)), reader.GetInt64());
 			reader = CreateReader(buffer);
@@ -216,8 +219,8 @@ namespace Com.Drew.Lang
 		{
 			long nanBits = unchecked((long)(0xfff0000000000001L));
 			Sharpen.Tests.IsTrue(double.IsNaN(Sharpen.Extensions.LongBitsToDouble(nanBits)));
-			sbyte[] buffer = new sbyte[] { unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xf0))), unchecked((int)(0x00)), unchecked((int)(0x00)), unchecked((int)(0x00)), unchecked((int)(0x00)), unchecked((int)(0x00)), unchecked(
-				(int)(0x01)) };
+			sbyte[] buffer = new sbyte[] { unchecked((sbyte)unchecked((int)(0xff))), unchecked((sbyte)unchecked((int)(0xf0))), unchecked((int)(0x00)), unchecked((int)(0x00)), unchecked((int)(0x00)), unchecked((int
+				)(0x00)), unchecked((int)(0x00)), unchecked((int)(0x01)) };
 			SequentialReader reader = CreateReader(buffer);
 			Sharpen.Tests.IsTrue(double.IsNaN(reader.GetDouble64()));
 		}
